Updated algorithm
