+++ src/ConsoleApp/Logic/IGameLogic.cs
@@ -1,9 +1,7 @@
-﻿using ConsoleApp.Models;
-
-namespace ConsoleApp.Logic;
+﻿namespace ConsoleApp.Logic;
 
 public interface IGameLogic
 {
-    IGame Play(int curNum);
-    IGame Start();
+    void Play();
+    void Start();
 }
